--- core/network/src/main/java/online/partyrun/partyrunapplication/core/network/MatchSourceManagerImpl.kt
@@ -1,73 +1,0 @@
-package online.partyrun.partyrunapplication.core.network
-
-import okhttp3.Response
-import okhttp3.sse.EventSource
-import okhttp3.sse.EventSourceListener
-import online.partyrun.partyrunapplication.core.common.network.MatchSourceManager
-import timber.log.Timber
-
-class MatchSourceManagerImpl : MatchSourceManager {
-    private lateinit var matchEventSource: EventSource
-    private lateinit var matchResultSource: EventSource
-
-    override fun createMatchEventSourceListener(
-        onEvent: (data: String) -> Unit,
-        onClosed: () -> Unit
-    ): EventSourceListener {
-        return createEventListener(onEvent, onClosed)
-    }
-
-    override fun connectMatchEventSource(eventSource: EventSource) {
-        matchEventSource = eventSource
-    }
-
-    override fun connectMatchResultEventSource(eventSource: EventSource) {
-        matchResultSource = eventSource
-    }
-
-    override fun disconnectMatchEventSource() {
-        if (::matchEventSource.isInitialized) {
-            matchEventSource.cancel()
-        }
-    }
-
-    override fun disconnectMatchResultEventSource() {
-        if (::matchResultSource.isInitialized) {
-            matchResultSource.cancel()
-        }
-    }
-
-    private fun createEventListener(
-        onEvent: (data: String) -> Unit,
-        onClosed: () -> Unit
-    ): EventSourceListener {
-        return object : EventSourceListener() {
-            override fun onOpen(eventSource: EventSource, response: Response) {
-                Timber.tag("Event").d("Connection Success")
-            }
-
-            override fun onClosed(eventSource: EventSource) {
-                Timber.tag("Event").d("Connection closed")
-                onClosed()
-            }
-
-            override fun onEvent(
-                eventSource: EventSource,
-                id: String?,
-                type: String?,
-                data: String,
-            ) {
-                onEvent(data)
-            }
-
-            override fun onFailure(
-                eventSource: EventSource,
-                t: Throwable?,
-                response: Response?
-            ) {
-                Timber.tag("Event").d("On Failure -: ${response?.body?.string()}")
-            }
-        }
-    }
-}
-
