+++ core/network/src/main/java/online/partyrun/partyrunapplication/core/network/datasource/MatchDataSource.kt
@@ -3,14 +3,21 @@ package online.partyrun.partyrunapplication.core.network.datasource
 import okhttp3.sse.EventSource
 import okhttp3.sse.EventSourceListener
 import online.partyrun.partyrunapplication.core.common.network.ApiResult
-import online.partyrun.partyrunapplication.core.model.match.MatchDecisionRequest
-import online.partyrun.partyrunapplication.core.model.match.MatchStatusResult
-import online.partyrun.partyrunapplication.core.model.match.UserSelectedMatchDistance
+import online.partyrun.partyrunapplication.core.network.model.request.MatchDecisionRequest
+import online.partyrun.partyrunapplication.core.network.model.response.MatchStatusResponse
+import online.partyrun.partyrunapplication.core.network.model.request.RunningDistanceRequest
 
 interface MatchDataSource {
-    fun connectEvent(url: String, listener: EventSourceListener): EventSource
 
-    suspend fun registerMatch(userSelectedMatchDistance: UserSelectedMatchDistance): ApiResult<MatchStatusResult>
-    suspend fun acceptMatch(matchDecisionRequest: MatchDecisionRequest): ApiResult<MatchStatusResult>
-    suspend fun declineMatch(matchDecisionRequest: MatchDecisionRequest): ApiResult<MatchStatusResult>
+    suspend fun registerMatch(runningDistanceRequest: RunningDistanceRequest): ApiResult<MatchStatusResponse>
+    suspend fun acceptMatch(matchDecisionRequest: MatchDecisionRequest): ApiResult<MatchStatusResponse>
+    suspend fun declineMatch(matchDecisionRequest: MatchDecisionRequest): ApiResult<MatchStatusResponse>
+
+    fun createMatchEventSourceListener(onEvent: (data: String) -> Unit, onClosed: () -> Unit): EventSourceListener
+    fun createEventSource(url: String, listener: EventSourceListener): EventSource
+    fun connectWaitingEventSource(eventSource: EventSource)
+    fun connectMatchResultEventSource(eventSource: EventSource)
+    fun disconnectWaitingEventSource()
+    fun disconnectMatchResultEventSource()
+
 }
